comp - Check..., skeleten loading
